api/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
coverage==7.2.7
Django==3.2
django-cors-headers==4.1.0
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.3.0
importlib-metadata==6.7.0
iniconfig==2.0.0
model-bakery==1.13.0
packaging==24.0
pluggy==1.2.0
PyJWT==2.8.0
pytest==7.4.4
pytest-cov==4.1.0
pytest-django==4.5.2
pytz==2025.2
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
zipp==3.15.0
